+++ src/core/io/rive/rive_format.cpp
@@ -1,156 +1,11 @@
 #include "rive_format.hpp"
 
-#include "rive_stream.hpp"
-#include "type_def.hpp"
+#include "rive_loader.hpp"
+
+#include <QJsonArray>
+#include <QJsonObject>
 
 glaxnimate::io::Autoreg<glaxnimate::io::rive::RiveFormat> glaxnimate::io::rive::RiveFormat::autoreg;
-
-namespace glaxnimate::io::rive {
-
-class RiveLoader
-{
-public:
-    RiveLoader(int vmaj, int vmin, model::Document* document, RiveStream& stream, RiveFormat* format)
-        : document(document),
-        stream(stream),
-        format(format)
-    {
-        Q_UNUSED(vmaj);
-        Q_UNUSED(vmin);
-    }
-
-    bool load_document()
-    {
-        extra_props = stream.read_property_table();
-
-        if ( stream.has_error() )
-        {
-            format->error(QObject::tr("Could not read property table"));
-            return false;
-        }
-
-        std::vector<Object> objects;
-
-        while ( !stream.has_error() && !stream.eof() )
-            objects.emplace_back(read_object());
-
-        return true;
-    }
-
-    bool gather_definitions(Identifier type_id, Object& object)
-    {
-        auto it = defined_objects.find(type_id);
-        if ( it == defined_objects.end() )
-        {
-            fail(QObject::tr("Unknown object of type %1").arg(type_id));
-            return {};
-        }
-
-        const auto& def = it->second;
-
-        object.definitions.push_back(&def);
-
-        if ( def.extends )
-        {
-            if ( !gather_definitions(def.extends, object) )
-                return false;
-        }
-
-        object.property_definitions.insert(def.properties.begin(), def.properties.end());
-        return true;
-    }
-
-    Object read_object()
-    {
-        auto type_id = stream.read_varuint();
-        if ( stream.has_error() )
-        {
-            fail(QObject::tr("Could not load object type ID"));
-            return {};
-        }
-
-        Object obj;
-        obj.definition_id = type_id;
-
-        if ( !gather_definitions(type_id, obj) )
-            return {};
-
-        while ( true )
-        {
-            Identifier prop_id = stream.read_varuint();
-            if ( stream.has_error() )
-            {
-                fail(QObject::tr("Could not load property ID in %1 (%2)").arg(type_id).arg(obj.definitions[0]->name));
-                return {};
-            }
-
-            if ( prop_id == 0 )
-                break;
-
-            auto prop_def = obj.property_definitions.find(prop_id);
-            if ( prop_def == obj.property_definitions.end() )
-            {
-                auto unknown_it = extra_props.find(prop_id);
-                if ( unknown_it == extra_props.end() )
-                {
-                    fail(QObject::tr("Unknown property %1 of %2 (%3)").arg(prop_id).arg(type_id).arg(obj.definitions[0]->name));
-                    return {};
-                }
-                else
-                {
-                    format->warning(QObject::tr("Skipping unknown property %1 of %2 (%3)").arg(prop_id).arg(type_id).arg(obj.definitions[0]->name));
-                }
-            }
-            else
-            {
-                obj.properties[prop_def->second.name] = read_property_value(prop_def->second.type);
-                if ( stream.has_error() )
-                {
-                    fail(QObject::tr("Error loading property %1 (%2) of %3 (%4)").arg(prop_id).arg(prop_def->second.name).arg(type_id).arg(obj.definitions[0]->name));
-                    return {};
-                }
-            }
-        }
-
-        qDebug() << obj.definitions[0]->name << obj.definition_id << obj.properties;
-        return obj;
-    }
-
-    void fail(const QString& message)
-    {
-        format->error(message);
-        failed = true;
-    }
-
-    QVariant read_property_value(PropertyType type)
-    {
-        switch ( type )
-        {
-            case PropertyType::Bool:
-                return bool(stream.next());
-            case PropertyType::Bytes:
-                return stream.read_raw_string();
-            case PropertyType::String:
-                return stream.read_string();
-            case PropertyType::VarUint:
-                return QVariant::fromValue(stream.read_varuint());
-            case PropertyType::Float:
-                return stream.read_float();
-            case PropertyType::Color:
-                return QColor::fromRgba(stream.read_uint());
-        }
-
-        return {};
-    }
-
-    model::Document* document;
-    RiveStream& stream;
-    RiveFormat* format;
-    PropertyTable extra_props;
-    bool failed = false;
-};
-
-} // namespace glaxnimate::io::rive
 
 
 bool glaxnimate::io::rive::RiveFormat::on_open(QIODevice& file, const QString&, model::Document* document, const QVariantMap&)
@@ -184,7 +39,7 @@ bool glaxnimate::io::rive::RiveFormat::on_open(QIODevice& file, const QString&, 
         return false;
     }
 
-    return RiveLoader(vmaj, vmin, document, stream, this).load_document();
+    return RiveLoader(stream, this).load_document(document);
 }
 
 bool glaxnimate::io::rive::RiveFormat::on_save(QIODevice&, const QString&, model::Document*, const QVariantMap&)
@@ -192,4 +47,52 @@ bool glaxnimate::io::rive::RiveFormat::on_save(QIODevice&, const QString&, model
     return false;
 }
 
+QJsonDocument glaxnimate::io::rive::RiveFormat::to_json(const QByteArray& binary_data)
+{
+    RiveStream stream(binary_data);
+    if ( stream.read(4) != "RIVE" )
+        return {};
 
+    auto vmaj = stream.read_varuint();
+    stream.read_varuint(); // version min
+    stream.read_varuint(); // file id
+
+    if ( stream.has_error() || vmaj != 7 )
+        return {};
+
+    QJsonArray objects;
+    for ( const auto& rive_obj : RiveLoader(stream, this).load_object_list() )
+    {
+        QJsonObject obj;
+
+        QJsonArray types;
+        for ( const auto& def : rive_obj.definitions )
+        {
+            QJsonObject jdef;
+            jdef["id"] = int(def->type_id);
+            jdef["name"] = def->name;
+            types.push_back(jdef);
+        }
+        obj["class"] = types;
+
+        QJsonArray props;
+        for ( const auto& p : rive_obj.property_definitions )
+        {
+            QJsonObject prop;
+            prop["id"] = int(p.first);
+            prop["name"] = p.second.name;
+            prop["type"] = int(p.second.type);
+            auto value = rive_obj.properties.value(p.second.name);
+            if ( value.userType() == QMetaType::QColor )
+                prop["value"] = value.value<QColor>().name();
+            else
+                prop["value"] = QJsonValue::fromVariant(value);
+            props.push_back(prop);
+        }
+        obj["properties"] = props;
+
+        objects.push_back(obj);
+    }
+
+    return QJsonDocument(objects);
+}